Report name of repeatedly registered histograms
Histograms should ONLY be listed once in the code
for a given process (the browser process?).  If 
two or more histograms have the same name, it is an
error, and a DCHECK will fire.  This code will also
display the name of the offending histogram.

r=sky
Review URL: http://codereview.chromium.org/16609

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5059e3a9309395612c024fba2294a5b4e02ccc38