Add user script support to extensions.
This is implemented mostly by relying on the existing
user script code. But since extension user scripts are
declared, not discovered in a directory, I had to add
support for adding 'lone' user scripts to
UserScriptMaster. This led to a bit of refactoring.

Note that this CL relies on:
http://codereview.chromium.org/18352

Review URL: http://codereview.chromium.org/18198

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bdbc87ca87fda52c3262240705d974030e9ed4b4