Add to SDCH histogramming
Define a histogram macro that is customizable, and precise, for detailed
examination of performance when needed.

Provide graceful degradation when entire SDCH window is not
received. We now blacklist the site with an exponential back-off.
This allows teh user to hit reload, and get not-SDCH content.

bug=1609306

r=huanr,mbelshe
Review URL: http://codereview.chromium.org/19718

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d43e0729bbe1028c64ec29387c5a214e5eaa6dc7