Hands off the intercept if 'unpatch' fails
If IATPatchFunction::Unpatch fails during RestoreImportedFunction
it means that we cannot safely unpatch the import address table 
patch. In this case its better to be hands off the intercept as 
trying to unpatch again in the destructor of IATPatchFunction is
not going to be any safer.

In real world, when we patch a plugin's SetCursor, we intercept
npswf.dll's IAT entry of SetCursor. It seems that our unpatch
fails when the plugin ref count goes to 0. It could be because
some one else has patched on top of us. Then, during CRT 
uninitialization at process shutdown, the destructor of 
IATPatchFunction is called. It detects that we haven't unpatched
yet and tries to unpatch. But at this time the plugin DLL is
unloaded and the IAT thunk is invalid. There's no point in 
trying to unpatch unloaded DLL's IAT :)


BUG=6886


Review URL: http://codereview.chromium.org/21044

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5e165473f8b211c86b352e69ac05265e3a066cc3