Linux: use opaque NativeViewIds
Currently we are still passing GtkWidget* into the renderer and
trusting the value on the way out. With this patch we switch to using
opaque values.

These opaque values are handled by a GtkNativeViewIdManger, a
singleton object which maintains the list of currently valid ids and
their current X window ids.

We don't pass the X window ids directly to the renderer because they
are a) guessable and b) possibly variable for a GtkWidget. From a
patch size point of view, the X window isn't current created at the
point where we need it so significant work would be needed to reorder
operations to fix that as well.

This patch also removes the GTK accesses from the BACKGROUND_X11
thread which were a temporary hack.

http://codereview.chromium.org/92110

BUG=9014,9869,10787


Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6639f5c6328b64ea800693b343d789b32c050cab