POSIX: Make --user-data-dir work when it doesn't exist.
Previously, --user-data-dir called PathService::Override which called
AbsolutePath. On POSIX, this calls realpath and fails if the given
path doesn't actually exist.

We need to do this on Windows because we change the current directory
in order to load plugins (at least). However, on POSIX we shouldn't
ever change cwd.

So, on POSIX, don't try to make the user-data-dir absolute.

TEST=On POSIX, start chrome with --user-data-dir=/tmp/xyz (where xyz is a directory which doesn't exist). Once running, check that /tmp/xyz exists and contains the expected files.


Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 07dc5ffd9a5d10d55e2299651fe7a826d7fd7aba