Implement temporary First Run Dialog on OS X
We use a modal dialog with a single checkbox on OS X.

We use the OSX defaults system since we want something quick and reliable.  The
dialog is displayed at a very early stage in Chrome startup (Before any
subsystems start relying on the stats variable).  This means there are a few
quirks in displaying the UI.

A change was also needed to our event handling code since when the dialog is
shown we spin an event loop at a very early stage in the process lifetime.

Changed default value for stats to false and updated unit tests to reflect that.

Also some misc. minor cleanup.

BUG=11971,12046

Review URL: http://codereview.chromium.org/115608

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1d8e4ce0743eed37a143d99e568c28abbe3202f1