Get correct request to finsh time for histogram
Corrected a typo in generating the stat, and evolved to a new histogram
name to avoid confusion.

Note that the time between start and finish still needs to be
fixed, as that part of the regression is not handled.

BUG=14678

r=davemoore
Review URL: http://codereview.chromium.org/132072

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 978077e09aef57e064f27b61808aeb3856c085a9