linux: unload plugin libraries after extracting mime types
I had not unloaded them as a premature optimization, thinking it'd be
better to leave 'em around since they'll eventually be loaded by a page.
But that happens in the plugin process and reading the mime type happens
in the browser process, so it didn't help anyway.

BUG=17444

Review URL: http://codereview.chromium.org/160303

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 803ef4ef5b14ff135b826425825291259f9ceeab