linux: call g_thread_init() at relevant startup points
According to glib docs, we need to do this if it's at all possible
for us to hit glib on multiple threads.  This may be happening
when we grab plugin metadata from the file thread.

(Note that this is *not* initializing the GDK locking system.)

BUG=18957

Review URL: http://codereview.chromium.org/174264

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b882adb77481606b12e71603a7f14a6dddcf2092